Start consolidating the module management code
